LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/type.exists-expected.txt

type.exists
HTMLCanvasElement is a property of window
Defined in "Web IDL" (draft)
Actual output:

PASS HTMLCanvasElement is a property of window
